Remove dropdown menu from radix + fix spoonacular +

+++ src/components/snap/ImageUpload/HeicDCode.worker.ts
@@ -5,7 +5,7 @@ const decoder = await init();
 const canvas = new OffscreenCanvas(0, 0);
 const ctx = canvas.getContext("2d")!;
 onmessage = async (e) => {
-  const { data, width, height } = decoder.decode(e.data);
+  const { data, width, height } = decoder.decode(e.data as Uint8Array);
   canvas.width = width;
   canvas.height = height;
   ctx.putImageData(new ImageData(data, width, height), 0, 0);
